+++ changes.md
@@ -8,3 +8,6 @@ Release/revision numbering is major only, no minors - i.e. 1,2,3... not 1.1, 1.2
 #### 2017-11-29 R7
 * Changed CT clamp to be enabled/disabled from command line.  
 * Various optimisations to cram program into ROM (mostly shortening strings).
+
+#### 2018-01-22 R8
+* Changed sscanf calls to use SCNu8 macro for uint_8t vars, SCNd8 macro for int_8t vars